refactor: define our own Link mark to rewrite hrefs
The short relative links used in markdown
need to be converted to working urls in the href attributes.

In order to properly serialize the content
the full urls in html need to be converted to relative links as well.

This commit splits out the href handling into a new helper.
This way it can be easily tested.

Signed-off-by: Azul <azul@riseup.net>